+++ frontend/app/src/entities/ipam/ip-namespaces/domain/get-ip-namespace-list.ts
