+++ frontend/app/src/entities/ipam/ip-namespaces/ui/ip-namespace-tabs.tsx
