+++ frontend/app/src/entities/ipam/ip-prefixes/ui/ip-prefix-details.tsx
