+++ frontend/app/src/entities/ipam/ip-prefixes/utils/get-ip-prefix-table-columns.tsx
