+++ frontend/app/src/entities/ipam/ip-prefixes/utils/get-prefix-attributes-visible-in-list-view.ts
