+++ frontend/app/src/entities/ipam/ipam-tree/ipam-tree.tsx
@@ -7,8 +7,8 @@ import { ITreeViewOnLoadDataProps, NodeId } from "react-accessible-treeview";
 import { Link, useNavigate, useParams } from "react-router";
 
 import { TREE_ROOT_ID } from "@/entities/ipam/constants";
+import { useCurrentIpNamespace } from "@/entities/ipam/ip-namespaces/ui/ip-namespace-provider";
 import { GET_PREFIXES_ONLY } from "@/entities/ipam/ipam-tree/api/prefixes";
-import { useCurrentIpNamespace } from "@/entities/ipam/namespaces/ui/ip-namespace-provider";
 import { getObjectDetailsUrl } from "@/entities/nodes/utils";
 import { genericSchemasAtom, nodeSchemasAtom } from "@/entities/schema/stores/schema.atom";
 import { Badge } from "@/shared/components/ui/badge";
