+++ frontend/app/src/pages/ipam/ipam-details-relationship-page.tsx
@@ -1,6 +1,6 @@
 import { IP_ADDRESS_GENERIC, IP_PREFIX_GENERIC } from "@/entities/ipam/constants";
 import { IpAddressManager } from "@/entities/ipam/ip-addresses/ui/ip-address-manager";
-import { IpPrefixManager } from "@/entities/ipam/prefixes/ui/ip-prefix-manager";
+import { IpPrefixManager } from "@/entities/ipam/ip-prefixes/ui/ip-prefix-manager";
 import { RelationshipTable } from "@/entities/nodes/relationships/ui/relationship-table/relationship-table";
 import { useSchema } from "@/entities/schema/ui/hooks/useSchema";
 import { isOfKind } from "@/entities/schema/utils/is-of-kind";
